kick pdf generation to its own package/file
